+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 
 ### Added
 
+- functionality to download and save FewsNet data
 - Pre-commit hook to check Sphinx documentation can be built
 - [`pip-compile`](https://github.com/jazzband/pip-tools#version-control-integration)
   pre-commit hook to update requirements files